Added inserting prcedures to some of archive tables. Added trigger that archives address after deleting it. Added trigger that archives courses after deleting it. Added trigger that archives student after deletion -> this one handles deletion and archiving student's address and all of his courses. Each trigger can handle multiple rows. Updated backup file.

+++ SchoolDatabase/Listeners/Tables/Courses.sql
@@ -1,6 +1,9 @@
-﻿CREATE TABLE [listeners].[Courses]
+﻿ 
+ 
+
+CREATE TABLE [listeners].[Courses]
 (
-	CourseId int IDENTITY (1,1),
+	CourseId bigint IDENTITY (1,1),
     CourseGrade decimal(2,1),
     DateOfAssessment DATE,
     CourseStudentId int NOT NULL,
